# Add balanced scoring middleware to improve client-side load-balancing based on server responses