--- a/conjure-go-client/httpclient/client.go
+++ b/conjure-go-client/httpclient/client.go
@@ -56,6 +56,7 @@ type clientImpl struct {
 	metricsMiddleware      Middleware
 
 	uris                          []string
+	uriScorer                     internal.BalancedURIScoringMiddleware
 	maxAttempts                   int
 	disableTraceHeaderPropagation bool
 	backoffOptions                []retry.Option
@@ -83,7 +84,7 @@ func (c *clientImpl) Delete(ctx context.Context, params ...RequestParam) (*http.
 }
 
 func (c *clientImpl) Do(ctx context.Context, params ...RequestParam) (*http.Response, error) {
-	uris := c.uris
+	uris := c.uriScorer.GetURIsInOrderOfIncreasingScore()
 	if len(uris) == 0 {
 		return nil, werror.ErrorWithContextParams(ctx, "no base URIs are configured")
 	}
@@ -161,6 +162,7 @@ func (c *clientImpl) doOnce(
 		// must precede the error decoders because they return a nil response and the metrics need the status code of
 		// the raw response.
 		c.metricsMiddleware,
+		c.uriScorer,
 		// must precede the client error decoder
 		b.errorDecoderMiddleware,
 		// must precede the body middleware so it can read the response body