--- a/conjure-go-client/httpclient/internal/balanced_scorer.go
+++ b/conjure-go-client/httpclient/internal/balanced_scorer.go
@@ -1,3 +1,17 @@
+// Copyright (c) 2021 Palantir Technologies. All rights reserved.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
 package internal
 
 import (
@@ -61,8 +75,8 @@ func (u *balancedScorer) GetURIsInOrderOfIncreasingScore() []string {
 }
 
 func (u *balancedScorer) RoundTrip(req *http.Request, next http.RoundTripper) (*http.Response, error) {
-	baseUri := getBaseURI(req.URL)
-	info, foundInfo := u.uriInfos[baseUri]
+	baseURI := getBaseURI(req.URL)
+	info, foundInfo := u.uriInfos[baseURI]
 	if foundInfo {
 		atomic.AddInt32(&info.inflight, 1)
 		defer atomic.AddInt32(&info.inflight, -1)