--- a/conjure-go-client/httpclient/internal/noop_scorer.go
+++ b/conjure-go-client/httpclient/internal/noop_scorer.go
@@ -1,0 +1,37 @@
+// Copyright (c) 2021 Palantir Technologies. All rights reserved.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package internal
+
+import (
+	"net/http"
+)
+
+type noopScorer struct {
+	uris []string
+}
+
+func (n *noopScorer) GetURIsInOrderOfIncreasingScore() []string {
+	return n.uris
+}
+
+func (n *noopScorer) RoundTrip(req *http.Request, next http.RoundTripper) (*http.Response, error) {
+	return next.RoundTrip(req)
+}
+
+func NewNoopURIScoringMiddleware(uris []string) URIScoringMiddleware {
+	return &noopScorer{
+		uris,
+	}
+}