## Before this PR
<!-- What's wrong with the current state of the world and why change it now? -->
CGR client does not account for server responses when load-balancing across multiple URIs and just randomizes order of URIs when retrying.

This can lead to undesirably retrying a URI many times even when the server is unavailable, such as during a node restart.

## After this PR
<!-- User-facing outcomes this PR delivers go below -->
Add middleware based on Dialogue's '[balanced' load balancing strategy](https://github.com/palantir/dialogue/blob/develop/dialogue-core/src/main/java/com/palantir/dialogue/core/BalancedScoreTracker.java) that uses an exponentially decaying reservoir to track recent response errors and prioritize URIs with fewer in-flight requests and recent failures.

Fixes https://github.com/palantir/conjure-go-runtime/issues/194

Benchmark results:
```
benchmark                                                                  old ns/op     new ns/op     delta
BenchmarkAllocWithBytesBufferPool/NoByteBufferPool/count=1-16              93798         98068         +4.55%
BenchmarkAllocWithBytesBufferPool/NoByteBufferPool/count=10-16             931710        991491        +6.42%
BenchmarkAllocWithBytesBufferPool/NoByteBufferPool/count=100-16            9291976       9621014       +3.54%
BenchmarkAllocWithBytesBufferPool/WithByteBufferPool/count=1-16            91936         96066         +4.49%
BenchmarkAllocWithBytesBufferPool/WithByteBufferPool/count=10-16           903859        977292        +8.12%
BenchmarkAllocWithBytesBufferPool/WithByteBufferPool/count=100-16          9042090       9767101       +8.02%
BenchmarkUnavailableURIs/OneAvailableServer/count=10-16                    949802        963606        +1.45%
BenchmarkUnavailableURIs/OneAvailableServer/count=100-16                   9191157       9566737       +4.09%
BenchmarkUnavailableURIs/OneAvailableServer/count=1000-16                  91739632      96296739      +4.97%
BenchmarkUnavailableURIs/FourAvailableServers/count=10-16                  943496        1009042       +6.95%
BenchmarkUnavailableURIs/FourAvailableServers/count=100-16                 9426424       9882186       +4.83%
BenchmarkUnavailableURIs/FourAvailableServers/count=1000-16                93641142      100149462     +6.95%
BenchmarkUnavailableURIs/OneOutOfFourUnavailableServers/count=10-16        1224338       1003200       -18.06%
BenchmarkUnavailableURIs/OneOutOfFourUnavailableServers/count=100-16       12311181      10127655      -17.74%
BenchmarkUnavailableURIs/OneOutOfFourUnavailableServers/count=1000-16      122187011     102564133     -16.06%
BenchmarkUnavailableURIs/OneOutOfThreeUnavailableServers/count=10-16       1310920       1015474       -22.54%
BenchmarkUnavailableURIs/OneOutOfThreeUnavailableServers/count=100-16      13159841      10126297      -23.05%
BenchmarkUnavailableURIs/OneOutOfThreeUnavailableServers/count=1000-16     130296842     100600932     -22.79%
BenchmarkUnavailableURIs/OneOutOfTwoUnavailableServers/count=10-16         1474337       993779        -32.59%
BenchmarkUnavailableURIs/OneOutOfTwoUnavailableServers/count=100-16        14900138      9999934       -32.89%
BenchmarkUnavailableURIs/OneOutOfTwoUnavailableServers/count=1000-16       147037762     100873436     -31.40%

benchmark                                                                  old allocs     new allocs     delta
BenchmarkAllocWithBytesBufferPool/NoByteBufferPool/count=1-16              105            105            +0.00%
BenchmarkAllocWithBytesBufferPool/NoByteBufferPool/count=10-16             1051           1051           +0.00%
BenchmarkAllocWithBytesBufferPool/NoByteBufferPool/count=100-16            10519          10517          -0.02%
BenchmarkAllocWithBytesBufferPool/WithByteBufferPool/count=1-16            104            104            +0.00%
BenchmarkAllocWithBytesBufferPool/WithByteBufferPool/count=10-16           1041           1041           +0.00%
BenchmarkAllocWithBytesBufferPool/WithByteBufferPool/count=100-16          10416          10416          +0.00%
BenchmarkUnavailableURIs/OneAvailableServer/count=10-16                    1051           1051           +0.00%
BenchmarkUnavailableURIs/OneAvailableServer/count=100-16                   10514          10516          +0.02%
BenchmarkUnavailableURIs/OneAvailableServer/count=1000-16                  105175         105167         -0.01%
BenchmarkUnavailableURIs/FourAvailableServers/count=10-16                  1051           1121           +6.66%
BenchmarkUnavailableURIs/FourAvailableServers/count=100-16                 10517          11215          +6.64%
BenchmarkUnavailableURIs/FourAvailableServers/count=1000-16                105194         112162         +6.62%
BenchmarkUnavailableURIs/OneOutOfFourUnavailableServers/count=10-16        1475           1121           -24.00%
BenchmarkUnavailableURIs/OneOutOfFourUnavailableServers/count=100-16       14868          11218          -24.55%
BenchmarkUnavailableURIs/OneOutOfFourUnavailableServers/count=1000-16      146090         112173         -23.22%
BenchmarkUnavailableURIs/OneOutOfThreeUnavailableServers/count=10-16       1611           1121           -30.42%
BenchmarkUnavailableURIs/OneOutOfThreeUnavailableServers/count=100-16      16192          11216          -30.73%
BenchmarkUnavailableURIs/OneOutOfThreeUnavailableServers/count=1000-16     160540         112181         -30.12%
BenchmarkUnavailableURIs/OneOutOfTwoUnavailableServers/count=10-16         1877           1121           -40.28%
BenchmarkUnavailableURIs/OneOutOfTwoUnavailableServers/count=100-16        18845          11217          -40.48%
BenchmarkUnavailableURIs/OneOutOfTwoUnavailableServers/count=1000-16       188889         112191         -40.60%

benchmark                                                                  old bytes     new bytes     delta
BenchmarkAllocWithBytesBufferPool/NoByteBufferPool/count=1-16              7990          8046          +0.70%
BenchmarkAllocWithBytesBufferPool/NoByteBufferPool/count=10-16             79878         80515         +0.80%
BenchmarkAllocWithBytesBufferPool/NoByteBufferPool/count=100-16            803411        806921        +0.44%
BenchmarkAllocWithBytesBufferPool/WithByteBufferPool/count=1-16            7898          7950          +0.66%
BenchmarkAllocWithBytesBufferPool/WithByteBufferPool/count=10-16           78979         79677         +0.88%
BenchmarkAllocWithBytesBufferPool/WithByteBufferPool/count=100-16          791049        795755        +0.59%
BenchmarkUnavailableURIs/OneAvailableServer/count=10-16                    79910         80433         +0.65%
BenchmarkUnavailableURIs/OneAvailableServer/count=100-16                   799244        806119        +0.86%
BenchmarkUnavailableURIs/OneAvailableServer/count=1000-16                  7998631       8057533       +0.74%
BenchmarkUnavailableURIs/FourAvailableServers/count=10-16                  80503         82500         +2.48%
BenchmarkUnavailableURIs/FourAvailableServers/count=100-16                 805206        824364        +2.38%
BenchmarkUnavailableURIs/FourAvailableServers/count=1000-16                8069882       8258605       +2.34%
BenchmarkUnavailableURIs/OneOutOfFourUnavailableServers/count=10-16        117464        82460         -29.80%
BenchmarkUnavailableURIs/OneOutOfFourUnavailableServers/count=100-16       1183451       825758        -30.22%
BenchmarkUnavailableURIs/OneOutOfFourUnavailableServers/count=1000-16      11607855      8265169       -28.80%
BenchmarkUnavailableURIs/OneOutOfThreeUnavailableServers/count=10-16       129511        82496         -36.30%
BenchmarkUnavailableURIs/OneOutOfThreeUnavailableServers/count=100-16      1299289       823909        -36.59%
BenchmarkUnavailableURIs/OneOutOfThreeUnavailableServers/count=1000-16     12871795      8266066       -35.78%
BenchmarkUnavailableURIs/OneOutOfTwoUnavailableServers/count=10-16         152420        82647         -45.78%
BenchmarkUnavailableURIs/OneOutOfTwoUnavailableServers/count=100-16        1529061       823380        -46.15%
BenchmarkUnavailableURIs/OneOutOfTwoUnavailableServers/count=1000-16       15334106      8261072       -46.13%
```

==COMMIT_MSG==
Add balanced scoring middleware to improve client-side load-balancing based on server responses
==COMMIT_MSG==

## Possible downsides?
<!-- Please describe any way users could be negatively affected by this PR. -->

Small performance regression (<5%) when client is used with a single URI

<!-- Reviewable:start -->
---
This change is [<img src="https://reviewable.io/review_button.svg" height="34" align="absmiddle" alt="Reviewable"/>](https://reviewable.io/reviews/palantir/conjure-go-runtime/208)
<!-- Reviewable:end -->