@bmoylan 2021-09-24
___
*[conjure-go-client/httpclient/client.go, line 164 at r4](https://reviewable.io/reviews/palantir/conjure-go-runtime/208#-MkNPuLk1anFcB86u_Q2:-MkNPuLk1anFcB86u_Q3:b-lxd7ph) ([raw file](https://github.com/palantir/conjure-go-runtime/blob/b7d9ae35b9a1f8878e580518f4c5d86b020a928d/conjure-go-client/httpclient/client.go#L164)):*
> ```Go
> 		// the raw response.
> 		c.metricsMiddleware,
> 		c.uriScorer,
> ```

maybe a comment over this saying it has to come before decoding errors so we can access the raw status code


<!-- Sent from Reviewable.io -->